n8n/agent_system/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
